# InfluencerWebExperience/openapi.yaml
openapi: 3.0.3
info:
  title: InfluencerCRB Web Experience Service API
  version: 1.0.0
  description: Brand authentication API for signup, login, logout, and Google/Facebook social auth.
servers:
  - url: https://localhost:8081
paths:
...
  /api/campaigns:
    get:
      summary: List campaigns
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "7f6d9c90-8f77-4f73-a8af-c4b5b2e5fabc"
                      userId: "0f88ea2e-4f11-4f30-b5a6-7d8e9f10a111"
                      name: "Summer Glow Launch"
                      budget: "24000"
                      status: "active"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "7f6d9c90-8f77-4f73-a8af-c4b5b2e5fabc"
                        userId: "0f88ea2e-4f11-4f30-b5a6-7d8e9f10a111"
                        name: "Summer Glow Launch"
                        budget: "24000"
                        status: "active"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
    post:
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/campaigns/{id}:
    get:
      summary: Get campaign by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update campaign
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete campaign
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /api/creators:
    get:
      summary: List creators
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Creator list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "d95d57d0-8be3-4a71-8d34-49ab8828a222"
                      userId: "0f88ea2e-4f11-4f30-b5a6-7d8e9f10a111"
                      name: "Lena Park"
                      handle: "@lenalights"
                      platform: "Instagram"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "d95d57d0-8be3-4a71-8d34-49ab8828a222"
                        userId: "0f88ea2e-4f11-4f30-b5a6-7d8e9f10a111"
                        name: "Lena Park"
                        handle: "@lenalights"
                        platform: "Instagram"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
    post:
      summary: Create creator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Created creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/creators/{id}:
    get:
      summary: Get creator by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update creator
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete creator
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /api/campaign-creators:
    get:
      summary: List campaign-creator assignments
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignId
          in: query
          schema:
            type: string
            format: uuid
        - name: creatorId
          in: query
          schema:
            type: string
            format: uuid
        - name: stage
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Assignment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "a89d2f61-37a3-4b48-95f8-06da45a0a333"
                      campaignId: "7f6d9c90-8f77-4f73-a8af-c4b5b2e5fabc"
                      creatorId: "d95d57d0-8be3-4a71-8d34-49ab8828a222"
                      stage: "outreach"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "a89d2f61-37a3-4b48-95f8-06da45a0a333"
                        campaignId: "7f6d9c90-8f77-4f73-a8af-c4b5b2e5fabc"
                        creatorId: "d95d57d0-8be3-4a71-8d34-49ab8828a222"
                        stage: "outreach"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
    post:
      summary: Create campaign-creator assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/campaign-creators/{id}:
    get:
      summary: Get campaign-creator assignment by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update campaign-creator assignment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete campaign-creator assignment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /api/campaign-creators/{id}/stage:
    patch:
      summary: Update campaign-creator stage
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      required: false
      description: 1-based page number. When omitted with size, defaults to 1.
      schema:
        type: integer
        minimum: 1
    SizeQuery:
      name: size
      in: query
      required: false
      description: Page size. When omitted with page, defaults to 20.
      schema:
        type: integer
        minimum: 1
...
  schemas:
    GenericJson:
      type: object
      additionalProperties: true
    PaginatedGenericJson:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GenericJson'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasPrevious:
          type: boolean
        hasNext:
          type: boolean
      required:
        - items
        - page
        - size
        - total
        - totalPages
        - hasPrevious
        - hasNext
    ListOrPaginatedGenericJson:
      oneOf:
        - type: array
          items:
            $ref: '#/components/schemas/GenericJson'
        - $ref: '#/components/schemas/PaginatedGenericJson'
...